requirements.txt
-i https://pypi.org/simple
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
bcrypt==4.1.2
cffi==1.16.0; platform_python_implementation != 'PyPy'
click==8.1.7; python_version >= '3.7'
cryptography==42.0.5
ecdsa==0.19.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
fastapi==0.110.1; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
idna==3.7; python_version >= '3.5'
passlib[bcrypt]==1.7.4
pyasn1==0.6.0; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.7.0; python_version >= '3.8'
pydantic-core==2.18.1; python_version >= '3.8'
pydantic-settings==2.2.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-jose[cryptography]==3.3.0
python-multipart==0.0.9; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
typing-extensions==4.11.0; python_version >= '3.8'
uvicorn==0.29.0; python_version >= '3.8'

fastapi~=0.110.1
pydantic~=2.7.0
python-dotenv~=1.0.1